DelegationActionBar tests
