+++ src/components/reusables/DelegationActionBar/DelegationActionBar.tsx
@@ -1,11 +1,4 @@
-import {
-  Accordion,
-  AccordionHeader,
-  AccordionContent,
-  ButtonVariant,
-  ButtonColor,
-  Button,
-} from '@digdir/design-system-react';
+import { ButtonVariant, ButtonColor, Button } from '@digdir/design-system-react';
 import { useState } from 'react';
 import { t } from 'i18next';
 import * as React from 'react';
