--- src/components/reusables/NewDelegationAccordion/index.ts
@@ -1,1 +1,0 @@
-export { NewDelegationAccordion, NewDelegationAccordionButtonType } from './NewDelegationAccordion';
