+++ src/components/reusables/index.ts
@@ -2,7 +2,6 @@ export { CompactDeletableListItem } from './CompactDeletableListItem';
 export { SummaryPage } from './SummaryPage';
 export { DeletableListItem } from './DeletableListItem';
 export { Line } from './Line';
-export { NewDelegationAccordion, NewDelegationAccordionButtonType } from './NewDelegationAccordion';
 export { PageContainer } from './PageContainer';
 export { NavigationButtons } from './NavigationButtons/NavigationButtons';
 export { UserInfoBar } from './UserInfoBar/UserInfoBar';
